some artist controller n repository fixes

+++ Modsenfy.PresentationLayer/Controllers/ArtistController.cs
@@ -23,6 +23,12 @@ namespace Modsenfy.PresentationLayer.Controllers
         public async Task<ActionResult<ArtistDto>> GetArtist(int id)
         {
             var artist = await _artistRepository.GetByIdWithJoins(id);
+
+            if (artist is null)
+            {
+                return BadRequest();
+            }
+
             var artistDto = _mapper.Map<ArtistDto>(artist);
 
             return Ok(artistDto);
@@ -31,7 +37,12 @@ namespace Modsenfy.PresentationLayer.Controllers
         [HttpPost]
         public async Task<ActionResult> CreateArtist(ArtistDto artistDto)
         {
-            var artist = new Artist();
+            var artist = _mapper.Map<Artist>(artistDto);
+
+            if (artist is null)
+            {
+                return BadRequest();
+            }
 
             await _artistRepository.Create(artist);
             await _artistRepository.SaveChanges();
@@ -43,6 +54,11 @@ namespace Modsenfy.PresentationLayer.Controllers
         public async Task<ActionResult> UpdateArtist(ArtistDto artistDto)
         {
             var artist = _mapper.Map<Artist>(artistDto);
+
+            if (artist is null)
+            {
+                return BadRequest();
+            }
             
             await _artistRepository.Update(artist);
             await _artistRepository.SaveChanges();
@@ -54,6 +70,11 @@ namespace Modsenfy.PresentationLayer.Controllers
         public async Task<ActionResult> DeleteArtist(int id)
         {
             var artist = await _artistRepository.GetById(id);
+
+            if (artist is null)
+            {
+                return BadRequest();
+            }
             
             _artistRepository.Delete(artist);
             await _artistRepository.SaveChanges();
@@ -65,12 +86,13 @@ namespace Modsenfy.PresentationLayer.Controllers
         public async Task<ActionResult<IEnumerable<Artist>>> GetSeveralArtists(List<int> ids)
         {
             var artists = await _artistRepository.GetSeveralArtists(ids);
-            List<ArtistDto> artistDtos = new List<ArtistDto>();
 
-            foreach (var artist in artists)
+            if (artists is null)
             {
-                artistDtos.Add(_mapper.Map<ArtistDto>(artist));
+                return BadRequest();
             }
+
+            IEnumerable<ArtistDto> artistDtos = _mapper.Map<IEnumerable<ArtistDto>>(artists);
             
             return Ok(artistDtos);
         }
@@ -79,12 +101,13 @@ namespace Modsenfy.PresentationLayer.Controllers
         public async Task<ActionResult<IEnumerable<AlbumDto>>> GetArtistAlbums(int id)
         {
             var albums = await _artistRepository.GetArtistAlbums(id);
-            List<AlbumDto> albumDtos = new List<AlbumDto>();
 
-            foreach (var album in albums)
+            if (albums is null)
             {
-                albumDtos.Add(_mapper.Map<AlbumDto>(album));
+                return BadRequest();
             }
+
+            IEnumerable<AlbumDto> albumDtos = _mapper.Map<IEnumerable<AlbumDto>>(albums);
 
             return Ok(albumDtos);
         }
@@ -93,12 +116,13 @@ namespace Modsenfy.PresentationLayer.Controllers
         public async Task<ActionResult<IEnumerable<TrackDto>>> GetArtistTracks(int id)
         {
             var tracks = await _artistRepository.GetArtistTracks(id);
-            List<TrackDto> trackDtos = new List<TrackDto>();
 
-            foreach (var track in tracks)
+            if (tracks is null)
             {
-                trackDtos.Add(_mapper.Map<TrackDto>(track));
+                return BadRequest();
             }
+
+            IEnumerable<TrackDto> trackDtos = _mapper.Map<IEnumerable<TrackDto>>(tracks);
 
             return Ok(trackDtos);
         }
@@ -107,12 +131,13 @@ namespace Modsenfy.PresentationLayer.Controllers
 		public async Task<ActionResult<IEnumerable<StreamDto>>> GetAlbumStreams(int id)
 		{
             var streams = await _artistRepository.GetArtistStreams(id);
-            List<StreamDto> streamDtos = new List<StreamDto>();
 
-            foreach (var stream in streams)
+            if (streams is null)
             {
-                streamDtos.Add(_mapper.Map<StreamDto>(stream));
+                return BadRequest();
             }
+
+            IEnumerable<StreamDto> streamDtos = _mapper.Map<IEnumerable<StreamDto>>(streams);
 
             return Ok(streamDtos);
 		}
